Added safetyChecks to setRelativePointAsImpassable method.
In the process of adding safetyChecks i noticed that there was some duplicate code when calculating a relative point.
So i created a private method calculateRelativePosition which is used by the methods setRelativePointAsImpassable and moveSensorCm.
In the Angle class i made the getters for angle and degree const methods,
so that i could give them as const reference parameters to calculateRelativePosition.

This method has been tested in main.cpp instead of test_main.cpp because the tests do not compile on this module.