Added a stack, tempNode struct and checkNodeMem function.
In this commit a stack has been implemented to keep track of branches/intersections. This will allow us to return to a branch after completing a direction. The tempNode struct is used to remember the information required to properly continue after returning to a branch/intersection. The checkNodeMem function will be used to detect infinite loops when nodes return to a previously visited node, thus forming a loop.