Added infinite loop detection.
Nodes pointing back to previous nodes caused infinite loops. This commit solves that problem by storing the key nodes in an array and checking if the node has already been processed. The stack is used to check if any key nodes are left to be explored, if not, the conversion has been completed.